Feature: add cleanup step for removing ignored folders or failed folders
By allowing for all failed items to be removed, the subset of the SAF batch of successful writes can still be used.
This is very helpful when the batch size is in the thousands.

This adds a failed status to the batch that designates if any given row is marked as having failed.
The ImportDataWriter has been broken up into an interface called ImportDataOperator allowing for the new cleanup class (ImportDataCleaner) to implement.

The new Cleanup button is only available when write has any failures or ignored failures.

+++ src/main/java/edu/tamu/di/SAFCreator/ImportDataWriter.java
@@ -8,67 +8,43 @@ import javax.swing.SwingWorker;
 
 import edu.tamu.di.SAFCreator.model.Batch;
 import edu.tamu.di.SAFCreator.model.FlagPanel;
+import edu.tamu.di.SAFCreator.model.ImportDataOperator;
 import edu.tamu.di.SAFCreator.model.Item;
 import edu.tamu.di.SAFCreator.model.Verifier;
 import edu.tamu.di.SAFCreator.model.Verifier.Problem;
 
-public class ImportDataWriter extends SwingWorker<Boolean, ImportDataWriter.WriterUpdates>
+public class ImportDataWriter extends SwingWorker<Boolean, ImportDataOperator.Updates> implements ImportDataOperator
 {
 	private Batch batch = null;
 	private JTextArea console = null;
 	private FlagPanel flags = null;
 
-	public class WriterUpdates {
-		private int processed;
-		private int total;
-
-		public WriterUpdates() {
-			processed = 0;
-			total = 0;
-		}
-
-		public WriterUpdates(int processed, int total) {
-			this.processed = processed;
-			this.total = total;
-		}
-
-		public int getProcessed() {
-			return processed;
-		}
-
-		public int getTotal() {
-			return total;
-		}
-
-		public void setProcessed(int processed) {
-			this.processed = processed;
-		}
-
-		public void setTotal(int total) {
-			this.total = total;
-		}
-	}
-
+	@Override
 	public Batch getBatch() {
 		return batch;
 	}
 
+	@Override
 	public void setBatch(Batch batch) {
 		this.batch = batch;
 	}
 
+	@Override
 	public void setConsole(JTextArea console) {
 		this.console = console;
 	}
 
+	@Override
 	public void setFlags(FlagPanel flags) {
 		this.flags = flags;
 	}
 
+	@Override
 	public JTextArea getConsole() {
 		return console;
 	}
 
+	@Override
 	public FlagPanel getFlags() {
 		return flags;
 	}
@@ -79,14 +55,19 @@ public class ImportDataWriter extends SwingWorker<Boolean, ImportDataWriter.Writ
 		boolean noErrors = true;
 		int itemCount = 1;
 		int totalItems = batch.getItems().size();
-		for(Item item : batch.getItems())
+
+		for (Item item : batch.getItems())
 		{
+			if (isCancelled()) {
+				break;
+			}
+
 			if (batch.isIgnoredRow(++itemCount)) {
 				File directory = new File(batch.getOutputSAFDir().getAbsolutePath() + File.separator + itemCount);
 				directory.delete();
 
 				console.append("\tSkipped item (row " + itemCount + "), because of verification failure.\n");
-				publish(new ImportDataWriter.WriterUpdates(itemCount-1, totalItems));
+				publish(new ImportDataOperator.Updates(itemCount-1, totalItems));
 				continue;
 			}
 
@@ -105,17 +86,18 @@ public class ImportDataWriter extends SwingWorker<Boolean, ImportDataWriter.Writ
 			}
 
 			if (hasError) {
+				batch.failedRow(itemCount);
 				console.append("\tFailed to write item (row " + itemCount + ") " + item.getSAFDirectory() + ".\n");
 			} else {
 				console.append("\tWrote item (row " + itemCount + ") " + item.getSAFDirectory() + ".\n");
 			}
 
-			if (isCancelled()) {
-				console.append("Cancelled writing SAF.\n");
-				return (Boolean) null;
-			}
+			publish(new ImportDataOperator.Updates(itemCount-1, totalItems));
+		}
 
-			publish(new ImportDataWriter.WriterUpdates(itemCount-1, totalItems));
+		if (isCancelled()) {
+			console.append("Cancelled writing SAF.\n");
+			return null;
 		}
 
 		console.append("Done writing SAF data.\n");
